backend/requirements.txt
fastapi>=0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
numpy>=1.26.2
tensorflow==2.15.0
pillow==10.1.0
python-dotenv==1.0.0
pydantic>=2.5.2
pydantic-settings>=2.1.0 